+++ changelog.md
@@ -12,6 +12,7 @@
   - `#hasBubbles` & `#hasSteam` in `tfc:spring` fluid type, same functionality handled by nullable parameter of `#bubbleParticle` & `#steamParticle`
 - Remove deprecated methods in custom climate model event
 - Remove old, non-functional chunk-data creation mechanism for custom climate models that was added in 1.2.6
+- Add support for Kube's ingredient and result modifiers to TFC's advanced crafting recipes
 
 ### 1.2.6
 - Add TFC Log block builder